Start to refactor reducer
