+++ src/reducer.js
@@ -2,6 +2,7 @@ import _ from 'lodash'
 import immutable from 'seamless-immutable'
 import {CHANGE_VALUE, VALIDATION_RESOLVED, CHANGE_MODEL} from './actions'
 import evaluateConditions from './evaluate-conditions'
+import {set, unset} from './immutable-utils'
 
 const INITIAL_VALUE = {
   errors: {},
@@ -12,73 +13,6 @@ const INITIAL_VALUE = {
 }
 export function initialState (state) {
   return _.defaults(state, INITIAL_VALUE)
-}
-
-export function set (item, path, value) {
-  const segments = path.split('.')
-  const segment = segments.shift()
-  const segmentIsArrayIndex = /^\d+$/.test(segment)
-
-  if (segmentIsArrayIndex) {
-    item = item || []
-    const index = parseInt(segment)
-
-    for (let i = 0; i < index + 1; i++) {
-      if (item.length < (i + 1)) {
-        item.concat(null)
-      }
-    }
-
-    const newValue = segments.length > 0 ? set(item[index], segments.join('.'), value) : value
-
-    // Return immutable array with item at index updated
-    return item.slice(0, index).concat(newValue).concat(item.slice(index + 1))
-  }
-
-  const object = item || immutable({})
-  const newValue = segments.length > 0 ? set(object[segment], segments.join('.'), value) : value
-
-  return object.set(segment, newValue)
-}
-
-/**
- * Unset value in an object given a path to the key to unset
- * @param {Immutable} obj - object to unset path within
- * @param {String} path - path to value to unset
- * @returns {Immutable} new value with path unset
- */
-export function unset (obj, path) {
-  const segments = path.split('.')
-
-  if (_.isArray(obj)) {
-    const key = segments.splice(0, 1)
-    const index = parseInt(key)
-
-    if (segments.length === 0) {
-      return obj.slice(0, index).concat(obj.slice(index + 1))
-    }
-
-    const newValue = unset(obj[index], segments.join('.'))
-
-    // NOTE: concatenating the newValue in an array to preserve multi-dimensional arrays
-    return obj.slice(0, index).concat([newValue]).concat(obj.slice(index + 1))
-  }
-
-  if (_.isObject(obj)) {
-    const key = segments.splice(0, 1)
-
-    if (segments.length === 0) {
-      return obj.without(key)
-    }
-
-    const newValue = unset(obj[key], segments.join('.'))
-    return obj.set(key, newValue)
-  }
-
-  // NOTE: explicitly checking for null because "typeof null === 'object'" which is misleading
-  const type = obj === null ? 'null' : typeof obj
-
-  throw new Error(`A path can only be unset for objects and arrays not ${type}`)
 }
 
 /**
@@ -103,64 +37,74 @@ function recursiveClean (value) {
   return output
 }
 
-export function reducer (state, action) {
-  switch (action.type) {
-    case CHANGE_VALUE:
-      const {value, bunsenId} = action
-      let newValue
+export const actionReducers = {
+  '@@redux/INIT': function (state, action) {
+    if (state && state.baseModel) {
+      let initialValue = state.value || {}
+      state.model = evaluateConditions(state.baseModel, recursiveClean(initialValue))
+      // leave this undefined to force consumers to go through the proper CHANGE_VALUE channel
+      // for value changes
+      state.value = undefined
+    }
 
-      if (bunsenId === null) {
-        newValue = immutable(recursiveClean(value))
+    const newState = initialState(state || {})
+    newState.value = immutable(newState.value)
+    return newState
+  },
+
+  [CHANGE_MODEL]: function (state, action) {
+    return _.defaults({
+      baseModel: action.model,
+      model: evaluateConditions(action.model, state.value)
+    }, state)
+  },
+
+  [CHANGE_VALUE]: function (state, action) {
+    const {value, bunsenId} = action
+    let newValue
+
+    if (bunsenId === null) {
+      newValue = immutable(recursiveClean(value))
+    } else {
+      newValue = immutable(state.value)
+
+      if (_.includes([null, ''], value) || (_.isArray(value) && value.length === 0)) {
+        newValue = unset(newValue, bunsenId)
       } else {
-        newValue = immutable(state.value)
-
-        if (_.includes([null, ''], value) || (_.isArray(value) && value.length === 0)) {
-          newValue = unset(newValue, bunsenId)
-        } else {
-          newValue = set(newValue, bunsenId, value)
-        }
+        newValue = set(newValue, bunsenId, value)
       }
-      const newModel = evaluateConditions(state.baseModel, newValue)
-      let model
-      if (!_.isEqual(state.model, newModel)) {
-        model = newModel
-      } else {
-        model = state.model
-      }
+    }
+    const newModel = evaluateConditions(state.baseModel, newValue)
+    let model
+    if (!_.isEqual(state.model, newModel)) {
+      model = newModel
+    } else {
+      model = state.model
+    }
 
-      return _.defaults({
-        value: immutable(newValue),
-        model
-      }, state)
+    return _.defaults({
+      value: immutable(newValue),
+      model
+    }, state)
+  },
 
-    case VALIDATION_RESOLVED:
-      return _.defaults({
-        validationResult: action.validationResult,
-        errors: action.errors
-      }, state)
-    case CHANGE_MODEL:
-
-      return _.defaults({
-        baseModel: action.model,
-        model: evaluateConditions(action.model, state.value)
-      }, state)
-    case '@@redux/INIT':
-      if (state && state.baseModel) {
-        let initialValue = state.value || {}
-        state.model = evaluateConditions(state.baseModel, recursiveClean(initialValue))
-        // leave this undefined to force consumers to go through the proper CHANGE_VALUE channel
-        // for value changes
-        state.value = undefined
-      }
-
-      const newState = initialState(state || {})
-      newState.value = immutable(newState.value)
-      return newState
-
-    default:
-      // TODO: allow consumer to pass in logger class other than console
-      console.error(`Do not recognize action ${action.type}`)
+  [VALIDATION_RESOLVED]: function (state, action) {
+    return _.defaults({
+      validationResult: action.validationResult,
+      errors: action.errors
+    }, state)
   }
+}
+
+export function reducer (state, action) {
+  if (action.type in actionReducers) {
+    const actionReducer = actionReducers[action.type]
+    return actionReducer(state, action)
+  }
+
+  // TODO: allow consumer to pass in logger class other than console
+  console.error(`Do not recognize action ${action.type}`)
+
   return state
 }
