+++ tests/reducer-test.js
@@ -1,19 +1,86 @@
 var expect = require('chai').expect
+var sinon = require('sinon')
 var reducerExports = require('../lib/reducer')
 var actions = require('../lib/actions')
-var immutable = require('seamless-immutable')
+var actionReducers = reducerExports.actionReducers
 var reducer = reducerExports.reducer
-var unset = reducerExports.unset
-
-function expectImmutable (actual, expected) {
-  if ('asMutable' in actual) {
-    actual = actual.asMutable()
-  }
-
-  expect(actual).to.eql(expected)
-}
 
 describe('reducer', function () {
+  var sandbox
+
+  beforeEach(function () {
+    sandbox = sinon.sandbox.create()
+    sandbox.stub(console, 'error')
+  })
+
+  afterEach(function () {
+    sandbox.restore()
+  })
+
+  ;[
+    '@@redux/INIT',
+    actions.CHANGE_MODEL,
+    actions.CHANGE_VALUE,
+    actions.VALIDATION_RESOLVED
+  ]
+    .forEach((actionType) => {
+      describe(`when action type is ${actionType}`, function () {
+        var action, reducedState, state
+
+        beforeEach(function () {
+          action = {
+            type: actionType
+          }
+
+          state = {
+            foo: 'bar'
+          }
+
+          function stub () {
+            return 'out'
+          }
+
+          sandbox.stub(actionReducers, actionType, stub)
+
+          reducedState = reducer(state, action)
+        })
+
+        it('calls correct action reducer', function () {
+          expect(actionReducers[actionType].callCount).to.equal(1)
+          expect(actionReducers[actionType].lastCall.args).to.eql([state, action])
+        })
+
+        it('returns reduced state from action reducer', function () {
+          expect(reducedState).to.equal('out')
+        })
+
+        it('does not log error', function () {
+          expect(console.error.callCount).to.equal(0)
+        })
+      })
+    })
+
+  describe('unknown state', function () {
+    var reducedState, state
+
+    beforeEach(function () {
+      state = {
+        foo: 'bar'
+      }
+
+      reducedState = reducer(state, {type: 'bolt'})
+    })
+
+    it('logs error', function () {
+      expect(console.error.callCount).to.equal(1)
+      expect(console.error.lastCall.args).to.eql(['Do not recognize action bolt'])
+    })
+
+    it('returns state that was passed in', function () {
+      expect(reducedState).to.equal(state)
+    })
+  })
+
   describe('initial state', function () {
     it('should be what we want', function () {
       var initialState = reducer({}, {type: '@@redux/INIT'})
@@ -139,604 +206,6 @@ describe('reducer', function () {
         validationResult: ['you look kinda fat'],
         value: {},
         baseModel: {}
-      })
-    })
-  })
-
-  describe('unset', function () {
-    it('throws error when trying to unset path on a boolean', function () {
-      expect(() => {
-        unset(true, 'foo')
-      }).to.throw('A path can only be unset for objects and arrays not boolean')
-    })
-
-    it('throws error when trying to unset path on an integer', function () {
-      expect(() => {
-        unset(1, 'foo')
-      }).to.throw('A path can only be unset for objects and arrays not number')
-    })
-
-    it('throws error when trying to unset path on a number', function () {
-      expect(() => {
-        unset(1.5, 'foo')
-      }).to.throw('A path can only be unset for objects and arrays not number')
-    })
-
-    it('throws error when trying to unset path on null', function () {
-      expect(() => {
-        unset(null, 'foo')
-      }).to.throw('A path can only be unset for objects and arrays not null')
-    })
-
-    it('throws error when trying to unset path on a string', function () {
-      expect(() => {
-        unset('test', 'foo')
-      }).to.throw('A path can only be unset for objects and arrays not string')
-    })
-
-    it('throws error when trying to unset path on undefined', function () {
-      expect(() => {
-        unset(undefined, 'foo')
-      }).to.throw('A path can only be unset for objects and arrays not undefined')
-    })
-
-    it('boolean', function () {
-      const obj = immutable({
-        foo: true
-      })
-      const actual = unset(obj, 'foo')
-      expectImmutable(actual, {})
-    })
-
-    it('integer', function () {
-      const obj = immutable({
-        foo: 1
-      })
-      const actual = unset(obj, 'foo')
-      expectImmutable(actual, {})
-    })
-
-    it('number', function () {
-      const obj = immutable({
-        foo: 1.5
-      })
-      const actual = unset(obj, 'foo')
-      expectImmutable(actual, {})
-    })
-
-    it('string', function () {
-      const obj = immutable({
-        foo: 'a'
-      })
-      const actual = unset(obj, 'foo')
-      expectImmutable(actual, {})
-    })
-
-    it('boolean within object', function () {
-      const obj = immutable({
-        foo: {
-          bar: true
-        }
-      })
-      const actual = unset(obj, 'foo.bar')
-      expectImmutable(actual, {foo: {}})
-    })
-
-    it('integer within object', function () {
-      const obj = immutable({
-        foo: {
-          bar: 1
-        }
-      })
-      const actual = unset(obj, 'foo.bar')
-      expectImmutable(actual, {foo: {}})
-    })
-
-    it('number within object', function () {
-      const obj = immutable({
-        foo: {
-          bar: 1.5
-        }
-      })
-      const actual = unset(obj, 'foo.bar')
-      expectImmutable(actual, {foo: {}})
-    })
-
-    it('string within object', function () {
-      const obj = immutable({
-        foo: {
-          bar: 'a'
-        }
-      })
-      const actual = unset(obj, 'foo.bar')
-      expectImmutable(actual, {foo: {}})
-    })
-
-    describe('boolean array', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: [true, false, true]
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.0')
-        expectImmutable(actual, {foo: [false, true]})
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.1')
-        expectImmutable(actual, {foo: [true, true]})
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.2')
-        expectImmutable(actual, {foo: [true, false]})
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo')
-        expectImmutable(actual, {})
-      })
-    })
-
-    describe('integer array', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: [1, 2, 3]
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.0')
-        expectImmutable(actual, {foo: [2, 3]})
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.1')
-        expectImmutable(actual, {foo: [1, 3]})
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.2')
-        expectImmutable(actual, {foo: [1, 2]})
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo')
-        expectImmutable(actual, {})
-      })
-    })
-
-    describe('number array', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: [1.5, 2.5, 3.5]
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.0')
-        expectImmutable(actual, {foo: [2.5, 3.5]})
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.1')
-        expectImmutable(actual, {foo: [1.5, 3.5]})
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.2')
-        expectImmutable(actual, {foo: [1.5, 2.5]})
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo')
-        expectImmutable(actual, {})
-      })
-    })
-
-    describe('string array', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: ['a', 'b', 'c']
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.0')
-        expectImmutable(actual, {foo: ['b', 'c']})
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.1')
-        expectImmutable(actual, {foo: ['a', 'c']})
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.2')
-        expectImmutable(actual, {foo: ['a', 'b']})
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo')
-        expectImmutable(actual, {})
-      })
-    })
-
-    describe('boolean array within array', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: [
-            [true, false, true]
-          ]
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.0.0')
-        expectImmutable(actual, {
-          foo: [
-            [false, true]
-          ]
-        })
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.0.1')
-        expectImmutable(actual, {
-          foo: [
-            [true, true]
-          ]
-        })
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.0.2')
-        expectImmutable(actual, {
-          foo: [
-            [true, false]
-          ]
-        })
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo.0')
-        expectImmutable(actual, {
-          foo: []
-        })
-      })
-    })
-
-    describe('integer array within array', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: [
-            [1, 2, 3]
-          ]
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.0.0')
-        expectImmutable(actual, {
-          foo: [
-            [2, 3]
-          ]
-        })
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.0.1')
-        expectImmutable(actual, {
-          foo: [
-            [1, 3]
-          ]
-        })
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.0.2')
-        expectImmutable(actual, {
-          foo: [
-            [1, 2]
-          ]
-        })
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo.0')
-        expectImmutable(actual, {
-          foo: []
-        })
-      })
-    })
-
-    describe('number array within array', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: [
-            [1.5, 2.5, 3.5]
-          ]
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.0.0')
-        expectImmutable(actual, {
-          foo: [
-            [2.5, 3.5]
-          ]
-        })
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.0.1')
-        expectImmutable(actual, {
-          foo: [
-            [1.5, 3.5]
-          ]
-        })
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.0.2')
-        expectImmutable(actual, {
-          foo: [
-            [1.5, 2.5]
-          ]
-        })
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo.0')
-        expectImmutable(actual, {
-          foo: []
-        })
-      })
-    })
-
-    describe('string array within array', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: [
-            ['a', 'b', 'c']
-          ]
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.0.0')
-        expectImmutable(actual, {
-          foo: [
-            ['b', 'c']
-          ]
-        })
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.0.1')
-        expectImmutable(actual, {
-          foo: [
-            ['a', 'c']
-          ]
-        })
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.0.2')
-        expectImmutable(actual, {
-          foo: [
-            ['a', 'b']
-          ]
-        })
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo.0')
-        expectImmutable(actual, {
-          foo: []
-        })
-      })
-    })
-
-    describe('boolean array within object', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: {
-            bar: [true, false, true]
-          }
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.bar.0')
-        expectImmutable(actual, {
-          foo: {
-            bar: [false, true]
-          }
-        })
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.bar.1')
-        expectImmutable(actual, {
-          foo: {
-            bar: [true, true]
-          }
-        })
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.bar.2')
-        expectImmutable(actual, {
-          foo: {
-            bar: [true, false]
-          }
-        })
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo.bar')
-        expectImmutable(actual, {
-          foo: {}
-        })
-      })
-    })
-
-    describe('integer array within object', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: {
-            bar: [1, 2, 3]
-          }
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.bar.0')
-        expectImmutable(actual, {
-          foo: {
-            bar: [2, 3]
-          }
-        })
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.bar.1')
-        expectImmutable(actual, {
-          foo: {
-            bar: [1, 3]
-          }
-        })
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.bar.2')
-        expectImmutable(actual, {
-          foo: {
-            bar: [1, 2]
-          }
-        })
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo.bar')
-        expectImmutable(actual, {
-          foo: {}
-        })
-      })
-    })
-
-    describe('number array within object', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: {
-            bar: [1.5, 2.5, 3.5]
-          }
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.bar.0')
-        expectImmutable(actual, {
-          foo: {
-            bar: [2.5, 3.5]
-          }
-        })
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.bar.1')
-        expectImmutable(actual, {
-          foo: {
-            bar: [1.5, 3.5]
-          }
-        })
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.bar.2')
-        expectImmutable(actual, {
-          foo: {
-            bar: [1.5, 2.5]
-          }
-        })
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo.bar')
-        expectImmutable(actual, {
-          foo: {}
-        })
-      })
-    })
-
-    describe('string array within object', function () {
-      var obj
-
-      beforeEach(function () {
-        obj = immutable({
-          foo: {
-            bar: ['a', 'b', 'c']
-          }
-        })
-      })
-
-      it('first item', function () {
-        const actual = unset(obj, 'foo.bar.0')
-        expectImmutable(actual, {
-          foo: {
-            bar: ['b', 'c']
-          }
-        })
-      })
-
-      it('middle item', function () {
-        const actual = unset(obj, 'foo.bar.1')
-        expectImmutable(actual, {
-          foo: {
-            bar: ['a', 'c']
-          }
-        })
-      })
-
-      it('last item', function () {
-        const actual = unset(obj, 'foo.bar.2')
-        expectImmutable(actual, {
-          foo: {
-            bar: ['a', 'b']
-          }
-        })
-      })
-
-      it('entire array', function () {
-        const actual = unset(obj, 'foo.bar')
-        expectImmutable(actual, {
-          foo: {}
-        })
       })
     })
   })
